Remove redundant config; try to force image pull
Cleaning things up a bit. For Tommy's previous commit, he
added a sleep command to the docker image (so we could see
it running on the cluster), and pushed it to dockerhub ...
but when I run this script, I don't get the updated image.

The nested docker config (which I've commented) is supposed
to make the 'docker run' command force a pull of an
updated version of the image, is '--pull always' is set ...
but it raises an error inside nextflow.  When I imanually
run the .command.sh command ... it works as it should.
More work is required here.